General development progress savepoint. Many small changes. NodeFactory appears to be working but there are some odd linking errors in some situations.
Node: A cc file is needed for otherwise pure inline classes with virtual members so that GCC knows where to place the vtable. At least one of the virtual members must have an implementation in that file. The implementation may be empty. Several cc files where added because of this.
Added QueueFactory to go along with the NodeFactory.