Saving current state of the files.
Testing has revealed a problem with the factory designs as
they have been put together. The plan was that the factories
would be static objects which would add themselves to a
mapping on construction. But, because there is no guarantee
in the order of static object initialization the mapping
may not be initialized before the static factory objects.