Removed a deadlock state from RemoteQueue and added more erro checking. It was possible for the RemoteQueue to get into a bad state where the file thread would be waiting for something to be read and the action thread would be waiting for an action but the queue had data in it that could be sent. Now under most conditions this wasn't a problem as the other side would likely block on reading at some point which would trigger the queue to send more. But this caused deadlock in some of the tests. So the possibility of being in this particular state was removed.
Also, checks for when thread creation in the tests where added. Running 1000
unit tests on hera in parallel causes some of them to error out with an error
of cannot create any more threads. If this error is not checked for then the
test waits for a thread which was never started.